requirements.txt
numpy
xlwt
openpyxl
xlsxwriter
requests
pyTelegramBotAPI
pandas
mysql-connector-python